+++ src/radical/utils/__init__.py
@@ -49,15 +49,17 @@ from .profile        import read_profiles, combine_profiles, clean_profile
 from .profile        import TIME, EVENT, COMP, TID, UID, STATE, MSG, ENTITY
 from .profile        import PROF_KEY_MAX
 
-# import utility methods
-from .ids            import *
-from .read_json      import *
-from .debug          import *
-from .misc           import *
-from .algorithms     import *
+from .json_io        import read_json, read_json_str, write_json
+from .json_io        import parse_json, parse_json_str
 from .which          import which
 from .tracer         import trace, untrace
 from .get_version    import get_version
+
+# import various utility methods
+from .ids            import *
+from .debug          import *
+from .misc           import *
+from .algorithms     import *
 
 # import decorators
 from .timing         import timed_method, epoch, dt_epoch
