+++ src/radical/utils/config.py
@@ -128,7 +128,7 @@ import munch
 
 from .misc       import find_module, is_string
 from .misc       import expand_env as ru_expand_env
-from .read_json  import read_json, write_json
+from .json_io    import read_json, write_json
 from .dict_mixin import dict_merge, DictMixin
 
 from .singleton  import Singleton
